pyproject.toml
[project]
name = "thesis-agent"
version = "0.1.0"
description = "Local, sandboxed thesis-writing agent built on deepagents + Karpathy LLM Wiki pattern."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "thesis-agent contributors" }]
keywords = ["thesis", "writing", "agent", "deepagents", "llm-wiki", "anthropic", "claude"]

dependencies = [
    "deepagents>=0.1.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.3.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "questionary>=2.0.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.40.0",
    "pypdf>=4.2.0",
    "python-docx>=1.1.0",
    "ebooklib>=0.18",
    "trafilatura>=1.12.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
thesis = "thesis_agent.cli:app"

[project.urls]
Homepage = "https://github.com/your-org/thesis-agent"
Repository = "https://github.com/your-org/thesis-agent"
Issues = "https://github.com/your-org/thesis-agent/issues"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/thesis_agent"]

[tool.hatch.build.targets.sdist]
include = [
    "src/thesis_agent",
    "skills",
    "examples",
    "AGENTS.md",
    "README.md",
    "LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
